pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["macbook <kriukov@pos-credir.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.7"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"
black = "^25.1.0"
flake8 = "^7.1.2"
mypy = "^1.15.0"
isort = "^6.0.1"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.0"
drf-spectacular = "^0.28.0"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.10"
django-cors-headers = "^4.7.0"
coverage = "^7.6.12"
ipython = "^9.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
